move LoadingButton to `platform`
