+++ src/applications/personalization/profile360/vet360/components/base/EditModal.jsx
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import Modal from '@department-of-veterans-affairs/formation-react/Modal';
 
 import Vet360EditModalErrorMessage from './EditModalErrorMessage';
-import LoadingButton from './LoadingButton';
+import LoadingButton from 'platform/site-wide/loading-button/LoadingButton';
 import Vet360EditModalActionButtons from './EditModalActionButtons';
 
 export default class Vet360EditModal extends React.Component {
